awards 컴포넌트 분리 및 award-footer 삭제
