+++ packages/footer/src/index.ts
@@ -1,6 +1,5 @@
 import DefaultFooter from './footers/default-footer'
 
 export { default as LogoFooter } from './footers/logo-footer'
-export { default as AwardFooter } from './footers/award-footer'
 
 export default DefaultFooter
